% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsavgmana.R
\name{hsavgmana}
\alias{hsavgmana}
\title{hsavgmana}
\usage{
hsavgmana(type)
}
\arguments{
\item{type}{Enter either "Deck" or "Card". When you enter "Deck", it will give the average mana cost for the deck you searched
by using hsdeck function in hstone package.\cr When you enter "Card", it will give the average mana cost for the cards
you searched by using hscard function in hstone package.}
}
\value{
The average mana cost.
}
\description{
This function shows the average mana for either cards you searched or the deck you searched. \cr
(You must first run either hscard or hsdeck function in hstone package to use this function)
}
\examples{
hsavgmana(type = "Deck")
hsavgmana(type = "Card")
}
