% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscard.R
\name{hscard}
\alias{hscard}
\title{hscard}
\usage{
hscard(
  id,
  secret,
  class = "",
  manaCost = "",
  attack = "",
  health = "",
  rarity = "",
  type = "",
  minionType = "",
  textFilter = "",
  orderBy = ""
)
}
\arguments{
\item{id}{A client Id you obtained from Blizzard api.}

\item{secret}{A client Secret you obtained from Blizzard api.}

\item{class}{Class classification you want to obtain. You must enter the slug of the class that can be obtained using hsmeta function.}

\item{manaCost}{Mana Cost value classification you want to obtain.}

\item{attack}{Attack value classification you want to obtain.}

\item{health}{Health value classification you want to obtain.}

\item{rarity}{Rarity classification you want to obtain. You must enter the slug of the class that can be obtained using hsmeta function.}

\item{type}{Type classification you want to obtain. You must enter the slug of the class that can be obtained using hsmeta function.}

\item{minionType}{Minion type classification you want to obtain. You must enter the slug of the class that can be obtained using hsmeta function.}

\item{textFilter}{Text classfication you want to obtain.}

\item{orderBy}{Order by either health, attack, or mana cost of the cards you searched.}
}
\value{
A data frame with cards based on your selected choice for each classifications.
}
\description{
This function shows the data frame with all the cards with your choice of classifications. For various search terms for each classification, \cr
you can use hsmeta function in hstone package.
}
\examples{
hscard(id = Sys.getenv("id"), secret = Sys.getenv("secret"), class = "mage", manaCost = 5, attack = 5, orderBy = "health")
hscard(id = Sys.getenv("id"), secret = Sys.getenv("secret"), rarity = "legendary", type = "spell", orderBy = "manaCost")
hscard(id = Sys.getenv("id"), secret = Sys.getenv("secret"), minionType = "murloc", textFilter = "Battlecry")
}
