% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscardused.R
\name{hscardused}
\alias{hscardused}
\title{hscardused}
\usage{
hscardused(id, secret, used = "")
}
\arguments{
\item{id}{A client Id you obtained from Blizzard api.}

\item{secret}{A client Secret you obtained from Blizzard api.}

\item{used}{Put either "high", "medium", or "low" in order to retrieve cards with high usage (4 and more times used), medium usage (2 to 3), and low usage (1 or less). \cr
Leave this blank to retrieve all the cards without such criteria.}
}
\value{
The data frame with cards and how many times they are used in top tier decks.
}
\description{
This function returns how many times the cards are used in each decks that are retrived from the topdeckcodes function in hstone package. \cr
(You must first run hstopdeckcodes function in hstone package to use this function)
}
\examples{
hscardused(id = Sys.getenv("id"), secret = Sys.getenv("secret"), used = "high")
hscardused(id = Sys.getenv("id"), secret = Sys.getenv("secret"))
}
