% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdeck.R
\name{hsdeck}
\alias{hsdeck}
\title{hsdeck}
\usage{
hsdeck(id, secret, deckcode)
}
\arguments{
\item{id}{A client Id you obtained from Blizzard api.}

\item{secret}{A client Secret you obtained from Blizzard api.}

\item{deckcode}{Enter the Hearthstone deckcode.}
}
\value{
All the cards that are in the deck.
}
\description{
This function shows all the cards that are in the deck you search.
}
\examples{
hsdeck(id = Sys.getenv("id"), secret = Sys.getenv("secret"), deckcode = "AAECAaoIAqH4AualAw7TAeAGkwn6qgL28AKP+wKc/wKMhQP2igO0lwPGmQP0mQPUpQPYqQMA")
}
