% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsexpectdeck.R
\name{hsexpectdeck}
\alias{hsexpectdeck}
\title{hsexpectdeck}
\usage{
hsexpectdeck()
}
\value{
Expectation of what the type of deck is.
}
\description{
This function expects what concept the deck is. Decks are divided into four categories: Highlander, Winnie, Aggro, and Control. \cr
(You must first run hsdeck function in hstone package to use this function)
}
\examples{
hsexpectdeck()
}
