% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmeta.R
\name{hsmeta}
\alias{hsmeta}
\title{hsmeta}
\usage{
hsmeta(id, secret, give)
}
\arguments{
\item{id}{A client Id you obtained from Blizzard api.}

\item{secret}{A client Secret you obtained from Blizzard api.}

\item{give}{Entering \code{"classes"}, \code{"types"}, \code{"minions"}, or \code{"rarities"} will give you the codes assigned for each classifications.}
}
\value{
A data frame with codes for that specific \code{give} you entered.
}
\description{
This function shows the data frame with codes for card's metadata.
For example, the codes specific for classifications such as classes, types, minions, and rarities.
}
\examples{
hsmeta(id = Sys.getenv("id"), secret = Sys.getenv("secret"), give = "classes")
hsmeta(id = Sys.getenv("id"), secret = Sys.getenv("secret"), give = "minions")
}
