% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstopdeckcodes.R
\name{hstopdeckcodes}
\alias{hstopdeckcodes}
\title{hstopdeckcodes}
\usage{
hstopdeckcodes()
}
\value{
Data frame with top tier decks and their codes.
}
\description{
This function shows the top tier decks in the current season as well as the deck codes for each decks.
}
\details{
The top tier decks and their codes are retrived from (http://metastats.net/). When the new season starts, the data may not be available
from few hours to a day or two.\cr Please, run this function again after a while when there are no decks yet available.
}
\examples{
hstopdeckcodes()
}
